requirements.txt
streamlit
ultralytics
opencv-python
yt-dlp
pyyaml
